requirements.txt
Django==4.2.7
django-colorfield==0.10.0
#django-rangefilter==0.11.0
django-admin-rangefilter
django-import-export==3.3.1
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
django-unfold
python-decouple==3.8
psycopg2-binary==2.9.9  
django-wpadmin
redis==5.0.1
Django==4.2.7
django-unfold==0.2.0
django-colorfield==0.10.0
#django-rangefilter==0.11.0
django-import-export==3.3.1
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
django-admin-interface==0.19.1
django-flat-responsive==2.0
django-admin-tools==0.9.0
django-grappelli==3.0.6
gunicorn==21.2.0

django-admin-volt
django-admin-interface
chartjs
django-jazzmin
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-yasg  # Pour la documentation Swagger